refactor: rename from s3 to aws
