+++ lib/workload/stateful/filemanager/filemanager-http-lambda/src/main.rs
@@ -4,9 +4,9 @@ use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
 use tracing_subscriber::{fmt, EnvFilter};
 
-use filemanager::database::s3::ingester::Ingester;
+use filemanager::database::aws::ingester::Ingester;
 use filemanager::database::Ingest;
-use filemanager::events::s3::collector_builder::CollecterBuilder;
+use filemanager::events::aws::collector_builder::CollecterBuilder;
 use filemanager::events::Collect;
 
 /// Handle SQS events by manually calling the SQS receive function. This is meant
