+++ lib/workload/stateful/filemanager/filemanager/src/clients/aws/config.rs
@@ -9,16 +9,18 @@ pub struct Config {
 
 #[automock]
 impl Config {
+    /// Create a new config.
+    pub fn new(inner: ConfigLoader) -> Self {
+        Self { inner }
+    }
+
     /// Load the config.
     pub async fn load(self) -> SdkConfig {
         self.inner.load().await
     }
-}
 
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            inner: defaults(BehaviorVersion::latest()),
-        }
+    /// Create a new config with default behaviour.
+    pub fn with_defaults() -> Self {
+        Self::new(defaults(BehaviorVersion::latest()))
     }
 }
