+++ lib/workload/stateful/filemanager/filemanager/src/clients/mod.rs
@@ -1,9 +1,1 @@
-//! This module provides wrappers around the AWS SDK which allows mocking them for tests.
-//!
-
-// Suppress warning when using mock_all
-#![allow(dead_code)]
-
-pub mod config;
-pub mod s3;
-pub mod sqs;
+pub mod aws;
