+++ lib/workload/stateful/filemanager/filemanager/src/lib.rs
@@ -1,4 +1,4 @@
-mod clients;
+pub mod clients;
 pub mod database;
 pub mod env;
 pub mod error;
